blog/requirements.txt
asgiref==3.4.1; python_version >= "3.8"
black==21.12b0; python_full_version >= "3.6.2"
click==8.0.3; python_version >= "3.6" and python_full_version >= "3.6.2"
colorama==0.4.4; python_version >= "3.6" and python_full_version >= "3.6.2" and platform_system == "Windows"
django==4.0; python_version >= "3.8"
djangorestframework-simplejwt==5.0.0; python_version >= "3.7"
djangorestframework==3.13.0; python_version >= "3.6"
flake8==4.0.1; python_version >= "3.6"
gunicorn==20.1.0; python_version >= "3.5"
mccabe==0.6.1; python_version >= "3.6"
mypy-extensions==0.4.3; python_full_version >= "3.6.2"
pathspec==0.9.0; python_full_version >= "3.6.2"
platformdirs==2.4.0; python_version >= "3.6" and python_full_version >= "3.6.2"
pycodestyle==2.8.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
pyflakes==2.4.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pyjwt==2.3.0; python_version >= "3.7"
pytz==2021.3; python_version >= "3.7"
sqlparse==0.4.2; python_version >= "3.8"
tomli==1.2.3; python_version >= "3.6" and python_full_version >= "3.6.2"
typing-extensions==4.0.1
tzdata==2021.5; sys_platform == "win32" and python_version >= "3.8"
whitenoise==5.3.0; python_version >= "3.5" and python_version < "4"
